A little more reorganizing for something sensible.
