+++ dynamo_consistency/backend/test.py
@@ -1,0 +1,12 @@
+"""
+This module defines dummy variables as needed by 
+dynamo_consistency.backend for running tests with
+"""
+
+# inventory
+# registry
+# siteinfo
+# get_listers
+# check_site
+# deletion_requests
+# DatasetFilter
