+++ dynamo_consistency/remotelister.py
@@ -5,25 +5,17 @@ Tool to get the files located at a site.
          Max Goncharov <maxi@mit.edu>
 """
 
+
 import logging
 
 from . import config
 from . import datatypes
 
-from .backend import listers
-from .backend import redirectors
-from .backend import siteinfo
-from .backend.cache import cache_tree
+from .cache import cache_tree
+from .backend import get_listers
 
 
 LOG = logging.getLogger(__name__)
-
-
-class NoPath(Exception):
-    """
-    An exception for when there is no known path to the remote site
-    """
-    pass
 
 
 @cache_tree('ListAge', 'remote')
@@ -32,70 +24,21 @@ def listing(site, callback=None):
     Get the information for a site, from XRootD or a cache.
 
     :param str site: The site name
-    :param function callback: The callback function to pass to :py:func:`datatypes.create_dirinfo`
+    :param function callback: The callback function to pass to
+                              :py:func:`datatypes.create_dirinfo`
     :returns: The site directory listing information
     :rtype: dynamo_consistency.datatypes.DirectoryInfo
-    :raises NoPath: When a way to list the remote site cannot be determined
     """
 
+    constructor, params = get_listers(site)
+
     config_dict = config.config_dict()
-    access = config_dict.get('AccessMethod', {})
-    if access.get(site) == 'SRM':
-        num_threads = int(config_dict.get('GFALThreads'))
-        LOG.info('threads = %i', num_threads)
-
-        backend = siteinfo.get_gfal_location(site)
-
-        if not backend:
-            raise NoPath('No path for gfal command for %s' % site)
-
-        directories = [
-            datatypes.create_dirinfo('/store', directory, listers.GFalLister,
-                                     [(site, backend, x) for x in xrange(num_threads)],
-                                     callback) \
-                for directory in config_dict.get('DirectoryList', [])
-        ]
-        # Return the DirectoryInfo
-        return datatypes.DirectoryInfo(name='/store', directories=directories)
-
-    # Get the redirector for a site
-    # The redirector can be used for a double check (not implemented yet...)
-    # The redir_list is used for the original listing
-    num_threads = int(config_dict.get('NumThreads'))
-
-    balancer, door_list = redirectors.get_redirector(site)
-    LOG.debug('Full redirector list: %s', door_list)
-
-    if site in config_dict.get('UseLoadBalancer', []) or \
-            (balancer and not door_list):
-        num_threads = 1
-        door_list = [balancer]
-
-    if not door_list:
-        raise NoPath('No doors found for %s' % site)
-
-    while num_threads > len(door_list):
-        if len(door_list) % 2:
-            door_list.extend(door_list)
-        else:
-            # If even number of redirectors and not using both, stagger them
-            door_list.extend(door_list[1:])
-            door_list.append(door_list[0])
-
-    # Strip off the extra threads
-    door_list = door_list[:num_threads]
-
-    # Create DirectoryInfo for each directory to search (set in configuration file)
-    # The search is done with XRootDLister objects that have two doors and the thread
-    # number as initialization arguments.
 
     directories = [
         datatypes.create_dirinfo(
-            '/store/', directory,
-            listers.XRootDSubShell if access.get(site) == 'directx' else listers.XRootDLister,
-            [(site, door, thread_num) for thread_num, door in enumerate(door_list)],
-            callback) for directory in config_dict.get('DirectoryList', [])
+            config_dict['RootPath'], directory,constructor, params, callback)
+        for directory in config_dict.get('DirectoryList', [])
         ]
 
     # Return the DirectoryInfo
-    return datatypes.DirectoryInfo(name='/store', directories=directories)
+    return datatypes.DirectoryInfo(name=config_dict['PathRoot'], directories=directories)
